Implement model load only option
After a certain time, the AWS Lambda function changes its status to
idle. It takes a while to wake up the function again, which results in
a timeout of the post request.
To circumvent this issue, this commit implements an option to only load
the pre-trained model without performing the inference. This then allows
asynchroneous calls to wake up the function while the user is making the
user inputs in the front-end application.