Try two API call attempts for prediction before throwing an error
The first prediction attempt might fail because the lambda function runs from
a cold start and the max. API response time (from the AWS side) is 29 seconds.
In such a scenario, the first attempt will cause the lambda function to
initialize. The second attempt will then be much faster and within the max.
API response time.